+++ examples/browser/webworkers/worker.js
@@ -1,4 +1,4 @@
-importScripts('https://unpkg.com/mathjs@6.6.4/dist/math.min.js')
+importScripts('https://unpkg.com/mathjs@6.6.5/dist/math.min.js')
 
 // create a parser
 const parser = self.math.parser()
@@ -19,7 +19,7 @@ self.addEventListener('message', function (event) {
   // build a response
   const response = {
     id: request.id,
-    result: result,
+    result: self.math.format(result),
     err: err
   }